pyproject.toml
[tool.poetry]
name = "tg-rag"
version = "0.1.0"
description = "Telegram bot with RAG"
authors = ["Jduun <isaevstanislav2003@mail.ru>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
litellm = "^1.57.8"
pydantic = "^2.10.5"
langchain = "^0.3.23"
langchain-openai = "^0.3.14"
langchain-community = "^0.3.21"
pdfminer-six = "^20250416"
faiss-cpu = "^1.10.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"
ruff = "^0.11.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
